Fix issues reported by SVACE tool in CKM module.
Reviewed-on: https://gerrit.iotivity.org/gerrit/5311
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
(cherry picked from commit 862f950c2548e138c2f2aab0678032ed8d21f4f5)
Reviewed-on: https://gerrit.iotivity.org/gerrit/5603
Tested-by: Randeep Singh <randeep.s@samsung.com>

+++ resource/csdk/security/provisioning/ck_manager/src/crl_generator.c
@@ -42,7 +42,7 @@ PKIError GenerateCRL (const UTF8String_t *issuerName,
     RelativeDistinguishedName_t *issuerRDN          = NULL;
     CertificateRevocationInfo_t *cri                = NULL;
 
-    uint32_t crlMaxSize = (CRL_MIN_SIZE +
+    uint32_t crlMaxSize = (uint32_t)(CRL_MIN_SIZE +
             numberOfRevoked * (sizeof(CertificateRevocationInfo_t) + 4));
 
     uint32_t i;
